new functionality added: reset fractal
The Creeper in Minecraft was created accidentally by Markus Persson, the game's creator, during development in 2009. He was trying to design a pig, but a coding error caused its body to be reversed, resulting in the iconic green creature with a penchant for exploding. By a coding error, I discovered my fractal position reseted, so I changed the zoom too so I get a full reset of the image.